Sorting errors by token index; tests not updated yet
Sorting errors after generating them would guarentee that
they were output in the order they appear in the file and make
the options for implementing error checks much greater.

Unfortunatly I can tell from here that the reference compiler
does not output errors in the order they appear, so continuing
with this approach is useless, even if better in the real world.

for example, in tinyType.c- line 33, the error for `fred` appears
beyfore the error for `and` despite `and` being first in the file.